update icon sets
