+++ package.json
@@ -34,6 +34,7 @@
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-vue": "^9.24.0",
     "msw": "^2.3.5",
+    "pinia-plugin-persistedstate": "^3.2.1",
     "postcss": "^8.4.38",
     "typescript": "^5.4.4"
   },
