+++ quasar.config.cjs
@@ -43,15 +43,15 @@ module.exports = configure(function (ctx) {
     // https://github.com/quasarframework/quasar/tree/dev/extras
     extras: [
       // 'ionicons-v4',
-      // 'mdi-v5',
+      'mdi-v7',
       // 'fontawesome-v6',
       // 'eva-icons',
       // 'themify',
       // 'line-awesome',
       // 'roboto-font-latin-ext', // this or either 'roboto-font', NEVER both!
 
-      'roboto-font', // optional, you are not bound to it
-      'material-symbols-rounded' // optional, you are not bound to it
+      'roboto-font' // optional, you are not bound to it
+      // 'material-symbols-rounded' // optional, you are not bound to it
     ],
 
     // Full list of options: https://v2.quasar.dev/quasar-cli-vite/quasar-config-js#build
@@ -125,7 +125,7 @@ module.exports = configure(function (ctx) {
         }
       },
 
-      iconSet: 'material-symbols-rounded', // Quasar icon set
+      iconSet: 'mdi-v7', // Quasar icon set
       lang: 'en-US', // Quasar language pack
 
       // For special cases outside of where the auto-import strategy can have an impact
@@ -136,7 +136,7 @@ module.exports = configure(function (ctx) {
       // directives: [],
 
       // Quasar plugins
-      plugins: ['Notify', 'Cookies', 'LocalStorage']
+      plugins: ['Notify', 'Cookies']
     },
 
     // animations: 'all', // --- includes all animations
