+++ src/api/privileges.ts
@@ -1,0 +1,5 @@
+import { api } from 'boot/axios'
+
+export const retrievePrivilegeTree = (username: string) => {
+  return api.get(`/privileges/${username}/tree`)
+}
