+++ src/mocks/common.ts
@@ -6,7 +6,7 @@ export const commonHandlers = [
 
     const username = info.get('username')
     // Read the intercepted request body.
-    return HttpResponse.json({ username }, {
+    return HttpResponse.json({ user: { username, avatar: '' }, access_token: 'sfa23asdfasdfasdf' }, {
       headers: {
         'Set-Cookie': 'logged_in=yes'
       }
