+++ src/mocks/privileges.ts
@@ -1,39 +1,250 @@
 import { http, HttpResponse } from 'msw'
-import type { Privilege } from 'src/models'
+import type { Privilege, PrivilegeTreeNode } from 'src/models'
 
-const datas: Privilege[] = []
-const subDatas: Privilege[] = []
+const datas: Privilege[] = [
+  {
+    id: 1,
+    path: '/system',
+    component: '#',
+    redirect: '/system/users',
+    name: 'system',
+    order: 1,
+    enabled: true,
+    icon: 'mdi-cog-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 7,
+    path: '/logs',
+    component: '#',
+    redirect: '/logs/operation',
+    name: 'logs',
+    order: 2,
+    enabled: true,
+    icon: 'mdi-clipboard-list-outline',
+    description: 'this is description for this row'
+  }
+]
 
-for (let i = 0; i < 20; i++) {
-  const data: Privilege = {
-    id: i,
-    name: 'privilege_' + i,
-    path: '/privilege_' + i,
-    icon: 'sym_r_home',
-    enabled: i % 3 > 0,
-    description: 'This is privilege description about xxx',
-    lastModifiedDate: new Date()
+const subDatas: Privilege[] = [
+  {
+    id: 2,
+    superiorId: 1,
+    path: 'groups',
+    component: 'pages/system/groups/IndexPage',
+    name: 'groups',
+    order: 1,
+    enabled: true,
+    icon: 'mdi-account-multiple-plus-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 3,
+    superiorId: 1,
+    path: 'users',
+    component: 'pages/system/users/IndexPage',
+    name: 'users',
+    order: 2,
+    enabled: true,
+    icon: 'mdi-account-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 4,
+    superiorId: 1,
+    path: 'privileges',
+    component: 'pages/system/privileges/IndexPage',
+    name: 'privileges',
+    order: 3,
+    enabled: true,
+    icon: 'mdi-shield-key-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 5,
+    superiorId: 1,
+    path: 'roles',
+    component: 'pages/system/roles/IndexPage',
+    name: 'roles',
+    order: 4,
+    enabled: true,
+    icon: 'mdi-shield-account-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 6,
+    superiorId: 1,
+    path: 'dictionaries',
+    component: 'pages/system/dictionaries/IndexPage',
+    name: 'dictionaries',
+    order: 5,
+    enabled: true,
+    icon: 'mdi-book-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 8,
+    superiorId: 7,
+    path: 'operation',
+    component: 'pages/logs/operation/IndexPage',
+    name: 'operationLog',
+    order: 1,
+    enabled: true,
+    icon: 'mdi-clipboard-text-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 9,
+    superiorId: 7,
+    path: 'access',
+    component: 'pages/logs/access/IndexPage',
+    name: 'accessLog',
+    order: 2,
+    enabled: true,
+    icon: 'mdi-file-document-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 10,
+    superiorId: 7,
+    path: 'audit',
+    component: 'pages/logs/audit/IndexPage',
+    name: 'auditLog',
+    order: 3,
+    enabled: true,
+    icon: 'mdi-clipboard-check-outline',
+    description: 'this is description for this row'
+  },
+  {
+    id: 11,
+    superiorId: 7,
+    path: 'scheduler',
+    component: 'pages/logs/scheduler/IndexPage',
+    name: 'schedulerLog',
+    order: 4,
+    enabled: true,
+    icon: 'mdi-calendar-text-outline',
+    description: 'this is description for this row'
   }
-  for (let j = 0; j < i; j++) {
-    const subData: Privilege = {
-      id: j,
-      name: 'privilege_' + i + '_' + j,
-      superiorId: i,
-      path: 'privilege_' + i + '_' + j,
-      icon: 'sym_r_group',
-      enabled: j % 2 > 0,
-      description: 'description',
-      lastModifiedDate: new Date()
-    }
-    subDatas.push(subData)
+]
+
+const treeNodes: PrivilegeTreeNode[] = [
+  {
+    id: 1,
+    path: '/system',
+    component: '#',
+    redirect: '/system/users',
+    name: 'system',
+    order: 1,
+    icon: 'mdi-cog-outline',
+    children: [
+      {
+        id: 2,
+        path: 'groups',
+        component: 'pages/system/groups/IndexPage',
+        name: 'groups',
+        order: 1,
+        icon: 'mdi-account-multiple-plus-outline'
+      },
+      {
+        id: 3,
+        path: 'users',
+        component: 'pages/system/users/IndexPage',
+        name: 'users',
+        order: 2,
+        icon: 'mdi-account-outline'
+      },
+      {
+        id: 4,
+        path: 'privileges',
+        component: 'pages/system/privileges/IndexPage',
+        name: 'privileges',
+        order: 3,
+        icon: 'mdi-shield-key-outline'
+      },
+      {
+        id: 5,
+        path: 'roles',
+        component: 'pages/system/roles/IndexPage',
+        name: 'roles',
+        order: 4,
+        icon: 'mdi-shield-account-outline'
+      },
+      {
+        id: 6,
+        path: 'dictionaries',
+        component: 'pages/system/dictionaries/IndexPage',
+        name: 'dictionaries',
+        order: 5,
+        icon: 'mdi-book-outline'
+      }
+    ]
+  },
+  {
+    id: 7,
+    path: '/logs',
+    component: '#',
+    redirect: '/logs/operation',
+    name: 'logs',
+    order: 2,
+    icon: 'mdi-clipboard-list-outline',
+    children: [
+      {
+        id: 8,
+        path: 'operation',
+        component: 'pages/logs/operation/IndexPage',
+        name: 'operationLog',
+        order: 1,
+        icon: 'mdi-clipboard-text-outline'
+      },
+      {
+        id: 9,
+        path: 'access',
+        component: 'pages/logs/access/IndexPage',
+        name: 'accessLog',
+        order: 2,
+        icon: 'mdi-file-document-outline'
+      },
+      {
+        id: 10,
+        path: 'audit',
+        component: 'pages/logs/audit/IndexPage',
+        name: 'auditLog',
+        order: 3,
+        icon: 'mdi-clipboard-check-outline'
+      },
+      {
+        id: 11,
+        path: 'scheduler',
+        component: 'pages/logs/scheduler/IndexPage',
+        name: 'schedulerLog',
+        order: 4,
+        icon: 'mdi-calendar-text-outline'
+      }
+    ]
   }
-  datas.push(data)
-}
+]
 
 export const privilegessHandlers = [
+  http.get('/api/privileges/:username/tree', ({ params }) => {
+    const { username } = params
+    console.log(username)
+    return HttpResponse.json(treeNodes)
+  }),
   http.get('/api/privileges/:id/subset', ({ params }) => {
-    const superiorId = params.id
-    return HttpResponse.json(subDatas.filter(item => item.superiorId === Number(superiorId)))
+    const { id } = params
+    return HttpResponse.json(subDatas.filter(item => item.superiorId === Number(id)))
+  }),
+  http.get('/api/privileges/:id', ({ params }) => {
+    const { id } = params
+    if (id) {
+      let res = datas.filter(item => item.id === Number(id))
+      if (!res) {
+        res = subDatas.filter(item => item.id === Number(id))
+      }
+      return HttpResponse.json(res)
+    }
+    return HttpResponse.json(null)
   }),
   http.get('/api/privileges', ({ request }) => {
     const url = new URL(request.url)
