+++ src/pages/system/dictionaries/IndexPage.vue
@@ -5,18 +5,18 @@
       <q-card style="min-width: 25em">
         <q-form @submit="onSubmit" @reset="onReset" class="q-gutter-md">
           <q-card-section>
-            <div class="text-h6">Role</div>
+            <div class="text-h6">Dictionary</div>
           </q-card-section>
 
           <q-card-section>
-            <q-input v-model="form.name" label="Role name" lazy-rules
+            <q-input v-model="form.name" label="Dictionary name" lazy-rules
               :rules="[val => val && val.length > 0 || 'Please type something']" />
 
-            <q-input v-model="form.description" label="Role deacription" type="textarea" />
+            <q-input v-model="form.description" label="Dictionary deacription" type="textarea" />
           </q-card-section>
 
           <q-card-actions align="right">
-            <q-btn title="cancel" type="reset" unelevated label="Cancel" v-close-popup />
+            <q-btn title="cancel" label="Cancel" unelevated type="reset" v-close-popup />
             <q-btn title="submit" type="submit" label="Submit" color="primary" />
           </q-card-actions>
 
@@ -24,18 +24,17 @@
       </q-card>
     </q-dialog>
 
-    <q-table flat ref="tableRef" :title="$t('roles')" selection="multiple" v-model:selected="selected" :rows="rows"
-      :columns="columns" row-key="id" v-model:pagination="pagination" :loading="loading" :filter="filter"
-      binary-state-sort @request="onRequest" class="full-width">
+    <q-table flat ref="tableRef" :title="$t('dictionaries')" :rows="rows" :columns="columns" row-key="id"
+      :loading="loading" v-model:pagination="pagination" binary-state-sort @request="onRequest" class="full-width">
       <template v-slot:top-right>
         <q-input dense debounce="300" v-model="filter" placeholder="Search">
           <template v-slot:append>
-            <q-icon name="sym_r_search" />
+            <q-icon name="mdi-search" />
           </template>
         </q-input>
-        <q-btn title="add" rounded color="primary" class="q-mx-md" :disable="loading" icon="sym_r_add"
-          :label="$t('add')" @click="addRow" />
-        <q-btn title="export" rounded outline color="primary" icon="sym_r_sim_card_download" :label="$t('export')"
+        <q-btn title="refresh" round flat color="primary" class="q-mx-md" :disable="loading" icon="mdi-refresh"
+          @click="refresh" />
+        <q-btn title="export" rounded outline color="primary" icon="mdi-file-download-outline" :label="$t('export')"
           @click="exportTable" />
       </template>
 
@@ -48,25 +47,28 @@
         </q-tr>
       </template>
 
-      <template v-slot:body-cell-members="props">
-        <q-td :props="props">
-          <q-avatar v-for="n in 5" :key="n" size="32px" :style="{ left: `${n * -2}px`, border: '2px solid white' }">
-            <img :src="`https://cdn.quasar.dev/img/avatar${n + 1}.jpg`" />
-          </q-avatar>
-        </q-td>
-      </template>
-      <template v-slot:body-cell-enabled="props">
-        <q-td :props="props">
-          <q-toggle v-model="props.row.enabled" size="sm" color="positive" />
-        </q-td>
-      </template>
-      <template v-slot:body-cell-id="props">
-        <q-td :props="props">
-          <q-btn title="edit" padding="xs" flat round color="primary" icon="sym_r_edit" @click="editRow(props.row.id)"
-            class="q-mt-none" />
-          <q-btn title="delete" padding="xs" flat round color="negative" icon="sym_r_delete"
-            @click="removeRow(props.row.id)" class="q-mt-none q-ml-sm" />
-        </q-td>
+      <template v-slot:body="props">
+        <q-tr :props="props">
+          <q-td auto-width>
+            <q-btn title="expand" round flat dense @click="props.expand = !props.expand"
+              :icon="props.expand ? 'mdi-chevron-down' : 'mdi-chevron-right'" />
+          </q-td>
+          <q-td v-for="col in props.cols" :key="col.name">
+            <div v-if="col.name === 'id'" class="text-right">
+              <q-btn title="edit" padding="xs" flat round color="primary" icon="mdi-pencil-outline"
+                @click="editRow(col.value)" class="q-mt-none" />
+            </div>
+            <div v-else-if="col.name === 'enabled'" class="text-center">
+              <q-toggle v-model="props.row.enabled" size="sm" color="positive" />
+            </div>
+            <span v-else>{{ col.value }}</span>
+          </q-td>
+        </q-tr>
+        <q-tr v-show="props.expand" :props="props">
+          <q-td colspan="100%" class="q-pr-none">
+            <sub-page v-if="props.expand" :title="props.row.name" :superior-id="props.row.id" />
+          </q-td>
+        </q-tr>
       </template>
     </q-table>
   </q-page>
@@ -77,9 +79,10 @@ import { ref, onMounted } from 'vue'
 import { exportFile, useQuasar } from 'quasar'
 import type { QTableProps } from 'quasar'
 import { api } from 'boot/axios'
+import SubPage from './SubPage.vue'
 
 import { SERVER_URL } from 'src/api/paths'
-import type { Role } from 'src/models'
+import type { Dictionary } from 'src/models'
 
 const $q = useQuasar()
 
@@ -88,9 +91,9 @@ const visible = ref<boolean>(false)
 const tableRef = ref()
 const rows = ref<QTableProps['rows']>([])
 const filter = ref('')
-const loading = ref<boolean>(false)
+const loading = ref(false)
 
-const form = ref<Role>({
+const form = ref<Dictionary>({
   name: '',
   description: ''
 })
@@ -103,17 +106,14 @@ const pagination = ref({
   rowsNumber: 0
 })
 
-const selected = ref([])
-
 const columns: QTableProps['columns'] = [
   { name: 'name', label: 'name', align: 'left', field: 'name', sortable: true },
-  { name: 'members', label: 'members', align: 'center', field: 'members' },
   { name: 'enabled', label: 'enabled', align: 'center', field: 'enabled' },
   { name: 'description', label: 'description', align: 'left', field: 'description' },
   { name: 'id', label: 'actions', field: 'id' }
 ]
 
-onMounted(() => {
+onMounted(async () => {
   tableRef.value.requestServerInteraction()
 })
 
@@ -128,7 +128,7 @@ async function onRequest(props: Parameters<NonNullable<QTableProps['onRequest']>
 
   const params = { page: page - 1, size: rowsPerPage, sortBy, descending, filter: filter || '' }
 
-  await api.get(SERVER_URL.ROLE, { params }).then(res => {
+  await api.get(SERVER_URL.DICTIONARY, { params }).then(res => {
     rows.value = res.data.content
     pagination.value.page = page
     pagination.value.sortBy = sortBy
@@ -141,13 +141,11 @@ async function onRequest(props: Parameters<NonNullable<QTableProps['onRequest']>
       message: error.message,
       type: 'negative'
     })
-  }).finally(() => {
-    loading.value = false
-  })
+  }).finally(() => { loading.value = false })
 }
 
-function addRow() {
-  visible.value = true
+function refresh() {
+  tableRef.value.requestServerInteraction()
 }
 
 function editRow(id: number) {
@@ -159,14 +157,6 @@ function editRow(id: number) {
       form.value = { ...row }
     }
   }
-}
-
-function removeRow(id: number) {
-  console.log('id: ', id)
-  loading.value = true
-  setTimeout(() => {
-    loading.value = false
-  }, 500)
 }
 
 function onSubmit() {
