+++ src/pages/system/privileges/SubPage.vue
@@ -1,9 +1,10 @@
 <template>
+
   <q-dialog v-model="visible" persistent>
     <q-card style="min-width: 25em">
       <q-form @submit="onSubmit" @reset="onReset" class="q-gutter-md">
         <q-card-section>
-          <div class="text-h6">Region</div>
+          <div class="text-h6">Group</div>
         </q-card-section>
 
         <q-card-section>
@@ -22,19 +23,15 @@
     </q-card>
   </q-dialog>
 
-  <q-table flat ref="tableRef" :title="title" selection="multiple" v-model:selected="selected" :rows="rows"
-    :columns="columns" row-key="id" v-model:pagination="pagination" :loading="loading" :filter="filter"
-    binary-state-sort @request="onRequest" class="full-width">
+  <q-table flat ref="tableRef" :title="title" :rows="rows" :columns="columns" row-key="id"
+    v-model:pagination="pagination" :loading="loading" :filter="filter" binary-state-sort @request="onRequest"
+    class="full-width">
     <template v-slot:top-right>
       <q-input dense debounce="300" v-model="filter" placeholder="Search">
         <template v-slot:append>
-          <q-icon name="sym_r_search" />
+          <q-icon name="mdi-search" />
         </template>
       </q-input>
-      <q-btn title="add" rounded color="primary" class="q-mx-md" :disable="loading" icon="sym_r_add" :label="$t('add')"
-        @click="addRow" />
-      <q-btn title="export" rounded outline color="primary" icon="sym_r_sim_card_download" :label="$t('export')"
-        @click="exportTable" />
     </template>
 
     <template v-slot:header="props">
@@ -50,17 +47,18 @@
       <q-tr :props="props">
         <q-td auto-width>
           <q-btn title="expand" round flat dense @click="props.expand = !props.expand"
-            :icon="props.expand ? 'sym_r_expand_less' : 'sym_r_expand_more'" />
+            :icon="props.expand ? 'mdi-chevron-down' : 'mdi-chevron-right'" />
         </q-td>
         <q-td v-for="col in props.cols" :key="col.name">
           <div v-if="col.name === 'id'" class="text-right">
-            <q-btn title="edit" padding="xs" flat round color="primary" icon="sym_r_edit" @click="editRow(props.row.id)"
-              class="q-mt-none" />
-            <q-btn title="delete" padding="xs" flat round color="negative" icon="sym_r_delete"
-              @click="removeRow(props.row.id)" class="q-mt-none q-ml-sm" />
+            <q-btn title="edit" padding="xs" flat round color="primary" icon="mdi-pencil-outline"
+              @click="editRow(col.value)" class="q-mt-none" />
           </div>
           <div v-else-if="col.name === 'enabled'" class="text-center">
             <q-toggle v-model="props.row.enabled" size="sm" color="positive" />
+          </div>
+          <div v-else-if="col.name === 'name'">
+            <q-icon :name="props.row.icon" size="sm" class="q-pr-sm" />{{ col.value }}
           </div>
           <span v-else>{{ col.value }}</span>
         </q-td>
@@ -77,15 +75,18 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue'
 import type { QTableProps } from 'quasar'
-import { exportFile, useQuasar } from 'quasar'
+import { useQuasar } from 'quasar'
 import { api } from 'boot/axios'
 
 import { SERVER_URL } from 'src/api/paths'
-import type { Region } from 'src/models'
+import type { Privilege } from 'src/models'
 
 const $q = useQuasar()
 
-const props_ = withDefaults(defineProps<{ title: string, superiorId?: number | undefined }>(), {
+const props = withDefaults(defineProps<{
+  title: string
+  superiorId?: number | undefined
+}>(), {
   title: '',
   superiorId: undefined
 })
@@ -97,10 +98,12 @@ const rows = ref<QTableProps['rows']>([])
 const filter = ref('')
 const loading = ref<boolean>(false)
 
-const form = ref<Region>({
+const form = ref<Privilege>({
   name: '',
-  areaCode: 0,
-  postalCode: 0,
+  path: '',
+  component: '',
+  icon: '',
+  order: 1,
   description: ''
 })
 
@@ -112,12 +115,9 @@ const pagination = ref({
   rowsNumber: 0
 })
 
-const selected = ref([])
-
 const columns: QTableProps['columns'] = [
   { name: 'name', label: 'name', align: 'left', field: 'name', sortable: true },
-  { name: 'postalCode', label: 'postalCode', align: 'left', field: 'postalCode', sortable: true },
-  { name: 'areaCode', label: 'areaCode', align: 'left', field: 'areaCode', sortable: true },
+  { name: 'path', label: 'path', align: 'left', field: 'path', sortable: true },
   { name: 'enabled', label: 'enabled', align: 'center', field: 'enabled' },
   { name: 'description', label: 'description', align: 'left', field: 'description' },
   { name: 'id', label: 'actions', field: 'id' }
@@ -130,22 +130,11 @@ onMounted(() => {
 /**
  * 查询列表
  */
-async function onRequest(props: Parameters<NonNullable<QTableProps['onRequest']>>[0]) {
+async function onRequest() {
   loading.value = true
 
-  const { page, rowsPerPage, sortBy, descending } = props.pagination
-  const filter = props.filter
-
-  const params = { page: page - 1, size: rowsPerPage, sortBy, descending, filter: filter || '' }
-
-  await api.get(`${SERVER_URL.REGION}/${props_.superiorId}/subset`, { params }).then(res => {
-    rows.value = res.data.content
-    pagination.value.page = page
-    pagination.value.sortBy = sortBy
-    pagination.value.rowsNumber = res.data.totalElements
-    pagination.value.rowsPerPage = rowsPerPage
-    pagination.value.sortBy = res.data.sortBy
-    pagination.value.descending = descending
+  await api.get(`${SERVER_URL.PRIVILEGE}/${props.superiorId}/subset`).then(res => {
+    rows.value = res.data
   }).catch(error => {
     $q.notify({
       message: error.message,
@@ -154,10 +143,6 @@ async function onRequest(props: Parameters<NonNullable<QTableProps['onRequest']>
   }).finally(() => {
     loading.value = false
   })
-}
-
-function addRow() {
-  visible.value = true
 }
 
 function editRow(id: number) {
@@ -171,58 +156,10 @@ function editRow(id: number) {
   }
 }
 
-function removeRow(id: number) {
-  console.log('id: ', id)
-  loading.value = true
-  setTimeout(() => {
-    loading.value = false
-  }, 500)
-}
-
 function onSubmit() {
   // Close the dialog after submitting
   visible.value = false
 }
 
 function onReset() { }
-
-function wrapCsvValue(val: string, formatFn?: (val: string, row?: string) => string, row?: string) {
-  let formatted = formatFn !== void 0 ? formatFn(val, row) : val
-
-  formatted = formatted === void 0 || formatted === null ? '' : String(formatted)
-
-  formatted = formatted.split('"').join('""')
-
-  return `"${formatted}"`
-}
-
-function exportTable() {
-  if (!columns || !rows.value || columns.length === 0 || rows.value.length === 0) {
-    // Handle the case where columns or rows are undefined or empty
-    console.error('Columns or rows are undefined or empty.')
-    return
-  }
-  // naive encoding to csv format
-  const content = [columns.map(col => wrapCsvValue(col.label))]
-    .concat(rows.value.map(row => columns.map(col =>
-      wrapCsvValue(typeof col.field === 'function' ? col.field(row) : row[col.field === void 0 ? col.name : col.field],
-        col.format,
-        row
-      )).join(','))
-    ).join('\r\n')
-
-  const status = exportFile(
-    'table-export.csv',
-    content,
-    'text/csv'
-  )
-
-  if (status !== true) {
-    $q.notify({
-      message: 'Browser denied file download...',
-      color: 'negative',
-      icon: 'warning'
-    })
-  }
-}
 </script>
