+++ src/stores/user-store.ts
@@ -1,31 +1,48 @@
 import { defineStore } from 'pinia'
-import { LocalStorage } from 'quasar'
+import { Cookies } from 'quasar'
+import { api } from 'boot/axios'
+import { retrievePrivilegeTree } from 'src/api/privileges'
+import type { Privilege } from 'src/models'
+import type { RouteRecordRaw } from 'vue-router'
 
 interface User {
-  username: string;
+  username: string
+  avatar: string
 }
 
 export const useUserStore = defineStore('user', {
   state: () => ({
-    user: null as User | null
+    user: null as User | null,
+    access_token: null as string | null,
+    privileges: [] as Privilege[],
+    routes: [] as RouteRecordRaw[]
   }),
-  getters: {
-    getUsername(): string | null {
-      const user = JSON.parse(LocalStorage.getItem('user') || '{}') as User | null
-      return this.user ? this.user.username : (user ? user.username : null)
+  actions: {
+    async logout() {
+      await api.post('/logout').then(() => {
+        Cookies.remove('logged_in')
+        this.$reset()
+      })
+    },
+    async login(username: string, password: string) {
+      await api.post('/login', new URLSearchParams({ username, password })).then(res => {
+        this.$patch({
+          user: res.data.user,
+          access_token: res.data.access_token
+        })
+        // privileges
+        retrievePrivilegeTree(username).then(response => {
+          this.$patch({
+            privileges: response.data
+          })
+        })
+      })
+    },
+    updateRoutes(routes: RouteRecordRaw[]) {
+      this.routes = routes
     }
   },
-  actions: {
-    updateUser(username: string) {
-      // 更新用户状态
-      this.user = { username }
-      LocalStorage.set('user', JSON.stringify(this.user))
-    },
-
-    clearUser() {
-      // 清除用户状态
-      this.user = null
-      LocalStorage.remove('user')
-    }
+  persist: {
+    paths: ['user', 'access_token', 'privileges']
   }
 })
